3DSOLID: corrections in member types of dxf_3dsolid_struct.

+++ src/3dsolid.h
@@ -1,7 +1,7 @@
 /*!
  * \file 3dsolid.h
  *
- * \author Copyright (C) 2012, 2013, 2014, 2015, 2016, 2017, 2018
+ * \author Copyright (C) 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019
  * by Bert Timmerman <bert.timmerman@xs4all.nl>.
  *
  * \brief Header file for a DXF 3D solid entity (\c 3DSOLID).
@@ -97,20 +97,20 @@ dxf_3dsolid_struct
                  * </ol>
                  * Group code = 60.\n
                  * \since Introduced in version R13. */
-        int color;
+        int16_t color;
                 /*!< Color of the entity.\n
                  * Defaults to \c BYLAYER if ommitted in the DXF file.\n
                  * Note that entities encapsulated in a block with the
                  * color \c BYBLOCK are represented in the "native" color of
                  * the \c BLOCK entity.\n
                  * Group code = 62. */
-        int paperspace;
+        int16_t paperspace;
                 /*!< Entities are to be drawn on either \c PAPERSPACE or
                  * \c MODELSPACE.\n
                  * Optional, defaults to \c DXF_MODELSPACE (0).\n
                  * Group code = 67.\n
                  * \since Introduced in version R13. */
-        int graphics_data_size;
+        int32_t graphics_data_size;
                 /*!< Number of bytes in the proxy entity graphics
                  * represented in the sub-sequent 310 groups, which are
                  * binary chunk records (optional).\n
@@ -199,7 +199,7 @@ dxf_3dsolid_struct
                  * group 1 string is greater than 255 characters
                  * (optional).\n
                  * Group code = 3. */
-        int modeler_format_version_number;
+        int16_t modeler_format_version_number;
                 /*!< Modeler format version number (currently = 1).\n
                  * Group code = 70. */
         char *history;
@@ -232,12 +232,12 @@ double dxf_3dsolid_get_linetype_scale (Dxf3dsolid *solid);
 Dxf3dsolid *dxf_3dsolid_set_linetype_scale (Dxf3dsolid *solid, double linetype_scale);
 int16_t dxf_3dsolid_get_visibility (Dxf3dsolid *solid);
 Dxf3dsolid *dxf_3dsolid_set_visibility (Dxf3dsolid *solid, int16_t visibility);
-int dxf_3dsolid_get_color (Dxf3dsolid *solid);
-Dxf3dsolid *dxf_3dsolid_set_color (Dxf3dsolid *solid, int color);
-int dxf_3dsolid_get_paperspace (Dxf3dsolid *solid);
-Dxf3dsolid *dxf_3dsolid_set_paperspace (Dxf3dsolid *solid, int paperspace);
-int dxf_3dsolid_get_graphics_data_size (Dxf3dsolid *solid);
-Dxf3dsolid *dxf_3dsolid_set_graphics_data_size (Dxf3dsolid *solid, int graphics_data_size);
+int16_t dxf_3dsolid_get_color (Dxf3dsolid *solid);
+Dxf3dsolid *dxf_3dsolid_set_color (Dxf3dsolid *solid, int16_t color);
+int16_t dxf_3dsolid_get_paperspace (Dxf3dsolid *solid);
+Dxf3dsolid *dxf_3dsolid_set_paperspace (Dxf3dsolid *solid, int16_t paperspace);
+int32_t dxf_3dsolid_get_graphics_data_size (Dxf3dsolid *solid);
+Dxf3dsolid *dxf_3dsolid_set_graphics_data_size (Dxf3dsolid *solid, int32_t graphics_data_size);
 int16_t dxf_3dsolid_get_shadow_mode (Dxf3dsolid *solid);
 Dxf3dsolid *dxf_3dsolid_set_shadow_mode (Dxf3dsolid *solid, int16_t shadow_mode);
 DxfBinaryGraphicsData *dxf_3dsolid_get_binary_graphics_data (Dxf3dsolid *solid);
@@ -264,8 +264,8 @@ DxfProprietaryData *dxf_3dsolid_get_proprietary_data (Dxf3dsolid *solid);
 Dxf3dsolid *dxf_3dsolid_set_proprietary_data (Dxf3dsolid *solid, DxfProprietaryData *proprietary_data);
 DxfProprietaryData *dxf_3dsolid_get_additional_proprietary_data (Dxf3dsolid *solid);
 Dxf3dsolid *dxf_3dsolid_set_additional_proprietary_data (Dxf3dsolid *solid, DxfProprietaryData *additional_proprietary_data);
-int dxf_3dsolid_get_modeler_format_version_number (Dxf3dsolid *solid);
-Dxf3dsolid *dxf_3dsolid_set_modeler_format_version_number (Dxf3dsolid *solid, int modeler_format_version_number);
+int16_t dxf_3dsolid_get_modeler_format_version_number (Dxf3dsolid *solid);
+Dxf3dsolid *dxf_3dsolid_set_modeler_format_version_number (Dxf3dsolid *solid, int16_t modeler_format_version_number);
 char *dxf_3dsolid_get_history (Dxf3dsolid *solid);
 Dxf3dsolid *dxf_3dsolid_set_history (Dxf3dsolid *solid, char *history);
 Dxf3dsolid *dxf_3dsolid_get_next (Dxf3dsolid *solid);
